+++ app/src/main/java/com/babiichuk/waterbalancetracker/app/ui/utils/Constants.kt
@@ -1,3 +1,3 @@
 package com.babiichuk.waterbalancetracker.app.ui.utils
 
-const val DEFAULT_WATER_RATE: Double = 2000.0
+const val DEFAULT_WATER_RATE: Int = 2000
