+++ app/src/main/java/com/babiichuk/waterbalancetracker/storage/entity/UserEntity.kt
@@ -12,15 +12,15 @@ data class UserEntity(
     val age: Int,
     val weight: Int,
     val gender: GenderType,
-    val recommendedWaterRate: Double,
-    val currentWaterRate: Double
+    val recommendedWaterRate: Int,
+    val currentWaterRate: Int
 ) {
 
     fun isNewUser(): Boolean{
         val ageEmpty = age == 0
         val weightEmpty = weight == 0
         val genderEmpty = gender == GenderType.UNKNOWN
-        val waterRateEmpty = recommendedWaterRate == 0.0
+        val waterRateEmpty = recommendedWaterRate == 0
         return ageEmpty || weightEmpty || genderEmpty || waterRateEmpty
     }
     companion object Factory {
@@ -31,8 +31,8 @@ data class UserEntity(
                 age = 0,
                 weight = 0,
                 gender = gender,
-                recommendedWaterRate = 0.0,
-                currentWaterRate = 0.0
+                recommendedWaterRate = 0,
+                currentWaterRate = 0
             )
         }
     }
